Replace annotator useEffect by 2 hooks: 1 for annotations/media and another for SAM; Move utils to a more appropriate folder
